rework io packages with io suffixes
Previously, there was a collision between package names for
zeek and zson, which created confusion. This commit unifies the
reader/writer implementations with an io suffic, e.g., zeekio,
ndjsonio, textio, etc
