+++ emitter/bytes.go
@@ -3,12 +3,12 @@ package emitter
 import (
 	"bytes"
 
-	"github.com/mccanne/zq/pkg/zsio"
-	"github.com/mccanne/zq/pkg/zsio/detector"
+	"github.com/mccanne/zq/pkg/zio"
+	"github.com/mccanne/zq/pkg/zio/detector"
 )
 
 type Bytes struct {
-	*zsio.Writer
+	*zio.Writer
 	buf bytes.Buffer
 }
 
@@ -16,7 +16,7 @@ func (b *Bytes) Bytes() []byte {
 	return b.buf.Bytes()
 }
 
-func NewBytes(format string, flags *zsio.Flags) (*Bytes, error) {
+func NewBytes(format string, flags *zio.Flags) (*Bytes, error) {
 	b := &Bytes{}
 	b.Writer = detector.LookupWriter(format, &noClose{&b.buf}, flags)
 	if b.Writer == nil {
