+++ pkg/zio/detector/recorder.go
