+++ pkg/zio/detector/track.go
